// src/abi/types/EnvoysAirdrop.ts
export type AirdropInfoStruct = {
  token: PromiseOrValue<string>;
  amount: PromiseOrValue<BigNumberish>;
  start: PromiseOrValue<BigNumberish>;
  end: PromiseOrValue<BigNumberish>;
};